# add on-the-fly-tabulation of calculated (eq/nonequilibrium) partition functions