## Description

choosing mode='tabulation' or (still default) mode='full summation' 'tabulation' is ~3500x faster at eq, ~500x faster at noneq (tested on CO2)

For a first working version:
- [x] on-the-fly tabulation implemented in PartitionFunction calculators  at equilibrium
- [x] on-the-fly tabulation implemented in PartitionFunction calculators  at non equilibrium (2T)
- [x] added unitary tests for PartitionFunctions
- [x] added a parameter, temporarily named `SpectrumFactory.misc.parsum_mode`
- [x] make it possible to compute Q without returnQvibQrot  in noneq mode 
- [x] compare (see Test below) tabulated parititon function & full summation on a noneq spectrum

For later: 
- [x] on-the-fly tabulation implemented in PartitionFunction calculators  at non equilibrium (3Tvib)
- [ ] implemnt on-the-fly tabulation with returnQvibQrot option
- [ ] implemnt on-the-fly tabulation with overpopulation option
- [ ] change default `parsum_mode` to 'tabulation' (wait for 0.9.31)
- [ ] mention on the docs in "Performance" section 

<!-- Please be sure to check out our contributing guidelines,
https://github.com/radis/radis/blob/develop/CONTRIBUTING.md . -->

## Test
<!-- Provide a general description of what your pull request does. -->

```python

from radis import SpectrumFactory

wmin, wmax = 2383.8, 2384.9
sf = SpectrumFactory(wavenum_min=wmin,
                     wavenum_max=wmax,
                      molecule='CO2',
                     isotope='1',
                     broadening_max_width=5,  # cm-1
                     medium='air',   
                     path_length=10.32, #cm
                     wstep=0.001, 
                     )
conditions = {'mole_fraction':0.2, 'pressure':1,
              'Ttrans':2000, 'Trot':2100, 'Tvib':1500}


sf.verbose = 0
sf.load_databank("HITEMP-CO2-DUNHAM") # my personal database
_  = sf.non_eq_spectrum(**conditions)  # initialize energies, etc.

# Now show times : 
sf.verbose = 3
sf.misc.parsum_mode = 'full summation'   # default
s_HITEMP = sf.non_eq_spectrum(**conditions)


# Now, again with tabulation:

sf.misc.parsum_mode = 'tabulation'
sf.verbose = 0
sf.load_databank("HITEMP-CO2-DUNHAM") # my personal database

_  = sf.non_eq_spectrum(**conditions)  # initialize energies, first tabulation

sf.verbose= 3
s_HITEMP2 = sf.non_eq_spectrum(**conditions)
```

Output : 
```
  warnings.warn(WarningType(message))
... 0.00s - Checked nonequilibrium parameters
... sorting lines by vibrational bands
... lines sorted in 0.1s
...... 0.60s - partition functions
...... 0.03s - populations
... 0.64s - Calculated nonequilibrium populations
...... 0.00s - map partition functions
...... 0.01s - corrected for populations and stimulated emission
... 0.02s - scaled nonequilibrium linestrength
... 0.03s - calculated emission integral
... 0.00s - Calculated lineshift
... 0.03s - Calculate broadening HWHM
... Calculating line broadening (4117 lines: expect ~ 2.06s on 1 CPU)
...... 0.01s - Precomputed DLM lineshapes (8)
...... 0.00s - Initialized vectors
...... 0.00s - Get closest matching line & fraction
...... 0.00s - Distribute lines over DLM
...... 0.01s - Convolve and sum on spectral range
...... 0.00s - Initialized vectors
...... 0.00s - Get closest matching line & fraction
...... 0.00s - Distribute lines over DLM
...... 0.02s - Convolve and sum on spectral range
... 0.07s - Calculated line broadening
... 0.01s - Calculated other spectral quantities
... 0.96s - Spectrum calculated (before object generation)
... 0.00s - Generated Spectrum object
0.96s - Spectrum calculated

# THEN BECOMES : 

... 0.00s - Checked nonequilibrium parameters
... sorting lines by vibrational bands
... lines sorted in 0.1s
...... 0.06s - partition functions
...... 0.01s - populations
... 0.08s - Calculated nonequilibrium populations
...... 0.00s - map partition functions
...... 0.00s - corrected for populations and stimulated emission
... 0.01s - scaled nonequilibrium linestrength
... 0.01s - calculated emission integral
... 0.00s - Calculated lineshift
... 0.02s - Calculate broadening HWHM
... Calculating line broadening (4117 lines: expect ~ 2.06s on 1 CPU)
...... 0.00s - Precomputed DLM lineshapes (8)
...... 0.00s - Initialized vectors
...... 0.00s - Get closest matching line & fraction
...... 0.00s - Distribute lines over DLM
...... 0.01s - Convolve and sum on spectral range
...... 0.00s - Initialized vectors
...... 0.00s - Get closest matching line & fraction
...... 0.00s - Distribute lines over DLM
...... 0.00s - Convolve and sum on spectral range
... 0.03s - Calculated line broadening
... 0.01s - Calculated other spectral quantities
... 0.22s - Spectrum calculated (before object generation)
... 0.00s - Generated Spectrum object
0.22s - Spectrum calculated


```
So about 4 faster for the total spectrum 


## Fixes 

A way to fix #45 without tabulating on-disk.