@erwanp 2021-07-14
@minouHub I did most of the implementaion on the Partition Function side, but there are things missing in the Factory to allow it to compute a noneq spectrum without computing the details of Qvib, dfQrot  (which current tabulated Partiont function implementatino doesn't do).   You did something similar in #312 so you could take over and implement it here. 

You may also want to discuss with @gagan-aryan as having a non-eq option without computing Qvib, dfQrot also reduces memory (no need to add [Qrot columns](https://github.com/radis/radis/pull/314/files#diff-300c4fe28c2e0881b2cca3e30a884a67ced2edf8a932b0c0aee094c32cc1c766L2355))

We do not ever need [nu_vib, nu_rot](https://github.com/radis/radis/blob/develop/radis/lbl/base.py#L2498) , etc. (this was discussed earlier but I can't find where?)  so all these columns could be removed. @minouHub no need to have a "if fast_parsum" mode as you did in #312, one mode where we don't compute the details of Qvib, dfQrot will be enough !  (faster, and less memory)