@gagan-aryan 2021-07-14
> We do not ever need nu_vib, nu_rot , etc. (this was discussed earlier but I can't find where?) so all these columns could be removed.

@erwanp You were discussing this with me. Below I quote a few messages from our discussion. 

> - For the partition functions : youll quickly find out that the nonequilibrium case is a bit more tricky than equilibrium. Find the lines where linestrength_noneq are corrected for lower state population. There is no Qgas/Qref ratio here because the lower state population can have an arbitrary shape. Instead we compute the populations nl directly, and use them
> - How does it apply to your optimization?
You may still do a map on Qref, easy.
But you can go further and look into where the populations nl are computed. Qgas is there too. But also some Qrot (depends not only on the isotope but also on the vibration group it belongs too)
> - Some of these are not necessarily needed to only get nl. Actually, there are used to get the fraction of the population in a rotational level nlrot,  or in a vib nlvib level. It could be discarded, or made optional. It's useful to plot the distributions but it cost you a few more columns. People who only want spectra don't care :)
> - All of that is very new. Actually, that's why Radis was developed initially, because there are no other infrared nonequilibrium codes available.
> - Short term : I think the Qref column  can be removed easily with a map(dict).
Then have a look at Qvib, Qrot, and see if some can be removed right away. Keep partial populations (nlvib, nlrot, etc.) in for the moment some people are using them for advanced cases ( @Minou).