@minouHub 2021-07-14
First of all: thank you Erwan for this contribution. To give context to @gagan-aryan (if needed): fitting spectra requires to simulate a bunch of them until the best solution is found. Thus, we need to calculate 100-1000 spectra for each measurement. Increasing the speed of the computation is therefore necessary.

I can implement what I've been doing on the partial sum for sure. Give me until Sunday to digest what you've done and push some stuff