@erwanp 2021-07-14
All good! I'm very confident it's robust and very fast, at least that's what the Proof of concept notebook showed. 

We also have to decide how to include it in the user interface. I think it should be a parameter when initializing the Factory. Currently I stored it as a misc.param (see the test above). Should be a computational parameter, as wstep for instance. 

@dcmvdbekerom you'll want to have a look because the Vaex(binby) approach seems very powerful for high dimension tabulation.