@minouHub 2021-07-19
There is a problem when vaex evaluate the mean. The following code

- does not work in spyder in **normal mode**, the code is stuck on `E_bins = df.mean("E", binby="logE", shape=N_bins)`
- work in spyder in **debugmode**
- work in command line `python my_code.py`

More puzzling, I have no problem using it in spyder with CO! I tried replacing the self.df['E'] values with random and it does not change the problem. So there is something with CO2, not CO, and it's not the energy values.

I don't understand what's going on. I tracked this hidden infinite loop and found in `vaex` that in the function `self._compute_agg`, the error happens when `self._delay(delay, var)` is called. 


```python
from radis import SpectrumFactory, plot_diff, get_residual
import matplotlib.pyplot as plt

# wmin, wmax = 2384, 2385
wmin, wmax = 2284, 2285
sf = SpectrumFactory(
    wavenum_min=wmin,
    wavenum_max=wmax,
    # molecule="CO",
    molecule="CO2",
    isotope="1",
    broadening_max_width=5,  # cm-1
    medium="air",
    path_length=10.32,  # cm
    wstep=0.001,
)
conditions = {
    "mole_fraction": 0.2,
    "pressure": 1,
    "Ttrans": 2000,
    "Trot": 2100,
    "Tvib": 1500,
}

# ... initialize tests :
sf.verbose = 0
sf.params.parsum_mode = "full summation"  # default
sf.load_databank("HITEMP-CO2-TEST")
# sf.load_databank('HITEMP-CO')
sf._add_bands()
_ = sf.non_eq_spectrum(**conditions)  # initialize energies, etc.
# ... compare:
sf.verbose = 3
s_HITEMP = sf.non_eq_spectrum(**conditions, name=sf.params.parsum_mode)

# Now, again with tabulation:
# ... init :
sf.params.parsum_mode = "tabulation"
sf.verbose = 0
sf.load_databank("HITEMP-CO2-TEST")
# sf.load_databank('HITEMP-CO')
sf._add_bands()
_ = sf.non_eq_spectrum(**conditions)  # initialize energies, first tabulation
# ... compare:
sf.verbose = 3
s_HITEMP2 = sf.non_eq_spectrum(**conditions, name=sf.params.parsum_mode)

plt.ion()
plot_diff(s_HITEMP2, s_HITEMP, method="ratio")

assert get_residual(s_HITEMP, s_HITEMP2, "abscoeff") < 1e-5
```