@minouHub 2021-07-19
Failed tests:
For `test_query.py`, the failure is due to HITRAN and should hopefully be sorted soon.

```python
ValueError: Error while parsing HITRAN output : <!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML 2.0//EN">
<html><head>
<title>503 Service Unavailable</title>
</head><body>
<h1>Service Unavailable</h1>
<p>The server is temporarily unable to service your
request due to maintenance downtime or capacity
problems. Please try again later.</p>
<hr>
<address>Apache/2.4.29 (Ubuntu) Server at hitran.org Port 443</address>
</body></html>
```