@minouHub 2021-07-20
@gagan-aryan, I've heard you know some stuff about vaex. Could you help with the problem we have above? The code ran in spyder (and only in spyder!) is usually stuck on `E_bins = df.mean("E", binby="logE", shape=N_bins)` and I cannot find why. I thought at some point it was due to the size of the df for CO2 but it is clearly not the case. I tried modify the df by the following expression and it clearly did not change anything. 
```python
keys = ["E", "grot", "gvib"]
df = vaex.from_pandas(self.df[keys][:10])
```
Also thought the two vaex call in `_eq_tabulation_setup ` and `_noneq_tabulation_setup `where messing with each others but again if I by-pass the first, the code is stuck on the second...

I even attempted a `conda update ananconda` without success ^^

Kind of stuck