@minouHub 2021-07-20
The problem was on my side and due to _something_ about my anaconda install. Thanks @erwanp and @gagan-aryan for your support. 
What are next steps? Should I work now (this weekend) on the future feature @erwanp requested or do you guys want to merge now?