@erwanp 2021-08-06
Latest changes aren't so conclusive; mainly because the default partition function improved in the meantime. 

On this example : 
```

from radis import SpectrumFactory

wmin, wmax = 2284, 2285
sf = SpectrumFactory(wavenum_min=wmin,
                     wavenum_max=wmax,
                      molecule='CO2',
                     isotope='1',
                     broadening_max_width=5,  # cm-1
                     medium='air',   
                     path_length=10.32, #cm
                     wstep=0.001, 
                     verbose=0,
                     )
conditions = {'mole_fraction':0.2, 'pressure':1,
              'Ttrans':2000, 'Trot':2100, 'Tvib':1500}

# Init:
sf.load_databank("HITEMP-CO2-TEST") # my personal database
_  = sf.non_eq_spectrum(**conditions)  # initialize energies, etc.

# Now show times : 
sf.params.parsum_mode = 'full summation'   # default
# sf.misc.parsum_mode = 'full summation'   # default
s_HITEMP = sf.non_eq_spectrum(**conditions)
s_HITEMP.print_perf_profile()


# Now, again with tabulation:

sf.params.parsum_mode = 'tabulation'
# sf.misc.parsum_mode = 'tabulation'
sf.load_databank("HITEMP-CO2-DUNHAM") # my personal database

_  = sf.non_eq_spectrum(**conditions)  # initialize energies, first tabulation

s_HITEMP2 = sf.non_eq_spectrum(**conditions)
s_HITEMP2.print_perf_profile()
```

I get : 

```
# BEFORE  
    spectrum_calculation      0.182s ████████████████
        check_line_databank             0.002s 
        check_non_eq_param              0.000s 
        reinitialize                    0.004s 
            copy_database                   0.001s 
            memory_usage_warning            0.003s 
            reset_population                0.000s 
        calc_noneq_population           0.111s █████████
            part_function                   0.022s █
            population                      0.089s ███████
        scaled_non_eq_linestrength      0.003s 
            map_part_func                   0.000s 
            corrected_population_se         0.003s 
        calc_emission_integral          0.011s 
        calc_lineshift                  0.001s 
        calc_hwhm                       0.012s █
        generate_wavenumber_arrays      0.000s 
        calc_line_broadening            0.034s ███
            precompute_DLM_lineshapes       0.001s 
            DLM_Initialized_vectors         0.000s 
            DLM_closest_matching_line       0.000s 
            DLM_Distribute_lines            0.001s 
            DLM_convolve                    0.030s ██
            others                          0.002s 
        calc_other_spectral_quan        0.003s 
        generate_spectrum_obj           0.000s 
        others                          0.002s


#  AFTER 
CO2spectrum2188947308096 profiler :
    spectrum_calculation      0.149s ████████████████
        check_line_databank             0.001s 
        check_non_eq_param              0.001s 
        reinitialize                    0.006s 
            copy_database                   0.003s 
            memory_usage_warning            0.003s 
            reset_population                0.000s 
        calc_noneq_population           0.071s ███████
            part_function                   0.065s ██████
            population                      0.006s 
        scaled_non_eq_linestrength      0.003s 
            map_part_func                   0.000s 
            corrected_population_se         0.003s 
        calc_emission_integral          0.012s █
        calc_lineshift                  0.000s 
        calc_hwhm                       0.014s █
        generate_wavenumber_arrays      0.000s 
        calc_line_broadening            0.038s ████
            precompute_DLM_lineshapes       0.001s 
            DLM_Initialized_vectors         0.000s 
            DLM_closest_matching_line       0.001s 
            DLM_Distribute_lines            0.001s 
            DLM_convolve                    0.034s ███
            others                          0.001s 
        calc_other_spectral_quan        0.003s 
        generate_spectrum_obj           0.000s 
        others                          0.001s 
```

so partition functions are not even 2x faster. 
And vaex remains stuck on Spyder console. 

However, if starting a fit with 200+ spectra, after the first being-stuck and slow start, vaex picks up and returns partiition functions within few ms (as in the initial results above).

Tried by reading vaex from a file on disk rather than with `vaex.from_pandas` ; same results.

Merging for further tests. The option is not on by default. Must be activated with: 
```parsum_mode='tabulation'```
t